requirements.txt
Flask==2.3.3
requests==2.31.0
google-generativeai==0.3.2
tushare==1.2.89
python-dotenv==1.0.0
markdown==3.5.1
pdfkit==1.0.0
weasyprint==61.2